requirements.txt
selenium
pytest
mysql-connector